Split logic into multiple files
