+++ constants.py
@@ -1,0 +1,44 @@
+EMPLOYEE = 1
+CUSTOMER = 2
+
+EMP_OPT_QUIT = 0
+EMP_OPT_ADD_FOOD_CATEGORY = 1
+EMP_OPT_ADD_FOOD_DETAILS = 2
+EMP_OPT_ADD_DELIVERY_PERSON = 3
+EMP_OPT_ASSIGN_DELIVERY_PERSON_TO_ORDER = 4 
+
+EMP_OPT_UPDATE_ORDER = 5
+EMP_OPT_UO_EN_ROUTE = 1
+EMP_OPT_UO_DELIVERED = 2
+
+EMP_OPT_VIEW_ORDER_DETAILS = 6
+EMP_OPT_VIEW_ORDER_STATUS = 7
+
+EMP_OPT_VIEW_REVENUE_TODAY = 8
+EMP_OPT_RT_CHECKEDOUT = 1
+EMP_OPT_RT_EN_ROUTE = 2
+
+EMP_OPT_DELETE_ORDER = 9
+
+CUST_OPT_QUIT = 0
+CUST_OPT_VIEW_MENU = 1
+CUST_OPT_CUSTOMER_SIGNUP = 2
+CUST_OPT_CUSTOMER_LOGIN = 3
+
+CUST_OPT_LOGOUT = 0
+
+CUST_OPT_PROCESS_ORDER = 1
+CUST_OPT_VIEW_ORDER = 2
+
+CUST_OPT_CHECKOUT = 3
+CUST_OPT_CHECKEDOUT = 1
+
+CUST_OPT_CANCEL_ORDER = 4
+CUST_OPT_CANCELLED = 1
+
+CUST_OPT_VIEW_ORDER_STATUS = 5
+
+CUST_OPT_BACK = 0
+CUST_OPT_ADD_FOOD_TO_ORDER = 1
+CUST_OPT_REMOVE_FOOD_TO_ORDER = 2
+CUST_OPT_UPDATE_FOOD_TO_ORDER = 3
